[LKW] Floating join points
This is the last known whereabouts of the floating join point. It took me far
too long to realize that join points really shouldn't be floating *at all*, and
anyway it's nigh impossible to get it to work right.

Essentially, floating join points work by splitting the floats in the
environment into two. Hence the arguments to the simplifier line up as

  F[E[J[e]]],

where F is the value floats, E is the continuation, J is the join floats, and
finally e is the expression.

So what happens when we need to add to the continuation? Suppose we have

  F[E[J[case e of A -> e_A; ... ]]].

Now, in order to rewrite this to

  F[E[J[(case [] of A -> e_A; ...)[e]]]],

we need to be able to add the join floats J to the continuation (so that the new
continuation is E[J[case [] of ...]]). But this is not what we want, since
case-of-case pushes the continuation inward and the floats are supposed to stay
floated *outward.*

Maybe mkDupableCont comes to the rescue? After all, its effect (in our regime
with floating join points) is to turn

  F[E[J[e]]]

into

  F[E_nodup[E_dup[J[e]]]],

where only E_dup gets pushed inward. If we make it so J as part of a
continuation can't be duplicated, everything stays floated.

The problem is, this procedure in using mkDupableCont, where part of the
continuation stays put, doesn't work with join points. (Here I'm describing
reallyRebuildCase, which splits the continuation, simplifies the alternatives
with the dupable part, then passes the whole rebuilt case to the non-dupable
part.) The issue is that, with join points as we know them, the simplifier
*cannot* decide to let part of the context sit there; it must *all* be pushed to
the leaves, so that it can be discarded at jumps. In other words,

  E[let join j = ... in case ... of A -> jump j 3; B -> f 4]

cannot be allowed to become

  let join j = E[...] in E[case ... of A -> jump j 3; B -> f 4]

because the context E has now been duplicated. (Also, the scope rules have been
violated - j is no longer tail-called relative to its own definition site). If
we push E *all the way* to the leaves, it's okay again:

  let join j = E[...] in case ... of A -> jump j 3; B -> E[f 4]

This is all to say that mkDupableCont now has to work a bit differently:

  E[e] => let join j = \x -> E_nodup[x] in (j [])[E_dup[e]]

But that only works if E_nodup doesn't contain join bindings that need to be
visible to E_dup. So you have to look for join floats in E_nodup and make sure
they float out along with j.

As things stand right now, things are rather jumbled. You have to be very
careful what ends up in the floats and what ends up in the continuation - right
now, in imaginary/paraffins, a join point winds up part of the continuation,
which floats inward right past its call sites. I'm not sure exactly what goes
wrong, but the code around mkDupableCont is kind of a mess anyway. mkDupableAlt
floats the join point it creates outward, while mkSimpleJoin adds it to the
continuation. Fixing that is nontrivial, because mkDupableAlt isn't returning a
continuation ...

However, what's dumb about all this is something I've only just managed to
realize:

  JOIN POINTS DON'T FLOAT.

Not in the simplifier, anyway. The places where a join point is allowed to be
invoked are

  1. in a case branch,
  2. in the definition of another join point, and
  3. in the body of a let.

But *nothing* floats out of a case branch, *nothing* floats out of a lambda (and
all join points are lambdas!), and we never move one binding past another
(because the one might refer to the other). So all of this is pointless. Well,
it's a convenience to be able to just throw things in the floats and let that
logic deal with what's a join point and what isn't; I'm not liking my ideas for
alternatives. Still. The idea of adding them to the continuation was always an
ugly, ugly hack, and that was the only way I could think of to keep things
tail-recursive. Not floating join points at all should clarify things.

(There is one place where we would want to float a join point: We currently
float them to top level, converting them to functions, given the opportunity,
since a top-level function isn't much worse than a continuation and now maybe
the residual is inlinable. But since nothing floats from a lambda, this could
only happen with a CAF, and seems unlikely that many CAFs have top-level join
points.)