New scheme for OccAnal join point analysis
We now keep in the UsageDetails three environments containing the ids that have
been zapped in various ways. This causes some overhead in needing to do extra
lookups and by wasting work combining OccAnals that will end up zapped, but the
big win is that now all the ids can be zapped at once simply by setting the
zapped set equal to the whole environment.

The new occurrence analyser still does more work than the original, enough to
require one test case to be adjusted, but the difference is now manageable.