Disallow eta-contracted join points again
The only concrete drawback to allowing join points to be eta-contracted appears
to be that an otherwise trivial join point might be chosen as a loop breaker.
This was easy enough to fix as a special case. (Most other uses of exprIsTrivial
are applied to arguments, not right-hand sides of local bindings.)

This reverts commit c68a0ebc8fdb265f002e0b35538ab238fa5e1303.