OccurAnal: Make loop-breaker scoring more uniform
Getting rid of an ugly hack meant to treat eta-contractible join points as
trivial. In actuality, *any* RHS small enough to be an eta-expanded trivial term
should have an unfolding with UnfWhen guidance, and hence it should be given a
very high score to keep it from being a loop breaker, but for some reason we
were only giving that score to *stable* UnfWhen unfoldings. Now we give *any*
UnfWhen unfolding a 6 and this catches the issue with join points and hopefully
more as well. I've flattened the cases to make more sensible decisions in other
ways, too; for instance, a stable unfolding could never get a 5 before.