Improve Core Lint code for join points and rules
We want to allow rules for join points to match against only a prefix of their
arguments, which is tricky when the rule's RHS is a jump. The old code blithely
assumed that the RHS of a rule for a join point would have a very particular
form; much better to search for that form intentionally and implement the
special case accordingly.