More sensible behavior with RULES, OccurAnal
Giving up on the idea of being flexible with how many arguments can be on the
LHS of a rule for a join point. Now it has to be exactly the join arity. But now
we do let the rhs of a join point rule have a tail call, so certain unfortunate
interactions are less likely (in particular, a SpecConstr-produced
specialization couldn't become a join point after its creation).