OccurAnal: Keep RHS context under big lambda
Deriving clauses for Generic produce a lot of RHSes of the form

  \(@x) -> C @x (f @x)

We don't want to inline f here any more than we would without the type
lambda---a constructor application as the RHS gets the arguments marked as
occurring inside a lambda, and the type lambda on the outside doesn't change
this. When I wrote occAnalLamOrRhs, I failed to notice the clause in occAnal
that keeps the RHS context when going under a type lambda. The result is mostly
a bunch of wasted effort as the inlined thing gets floated out again, but it
could get pretty bad (a ~25% increase in allocs in perf/compiler/T5642, for
instance).