Eliminate mkParallelBindings in favor of type rule for beta redexes
mkParallelBindings was introduced as an alternative to beta redexes like

(\@x y z -> e) @a e1 e2

where we don't want to traverse e again to perform the type substitution. The
tricky part is that naively let-binding everything winds up capturing any free
occurrences of y in e2. So the beta redex is easier to produce. Unfortunately,
if e contains a jump, the beta redex is ill typed. But since a beta redex is
*basically* just a let, we can introduce an exception to the typing rules to
allow a jump from the body of a beta redex.