Allow high-arity join points again
For eta-expansion purposes, we definitely want to allow the arity of a join
point to be high; this doesn't quite comport with the definition of a function's
arity given in CoreArity but it allows more eta-expansion for the surrounding
expression. We do have to be sure and get the arity exactly right in Core Prep,
however.